refactor(licensing): consolidate duplicate signLicense helper
Remove test-only signLicense from testing/keypair.ts now that the
production signLicense helper exists. verify-license.spec.ts now
tests against the real signer.
