+++ libs/licensing/src/testing.ts
@@ -2,7 +2,8 @@
 // Monorepo-internal test helpers. NOT part of the published package —
 // excluded from `tsconfig.lib.json` so nothing here ships in dist.
 // Downstream consumers cannot import `@cacheplane/licensing/testing`.
-export { generateKeyPair, signLicense } from './lib/testing/keypair';
+export { generateKeyPair } from './lib/testing/keypair';
 export type { DevKeyPair } from './lib/testing/keypair';
+export { signLicense } from './lib/sign-license';
 export { __resetRunLicenseCheckStateForTests } from './lib/run-license-check';
 export { __resetNagStateForTests } from './lib/nag';
